flask_service/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-RESTful>=0.3.10
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
requests>=2.31.0
gunicorn>=21.2.0

# Optional: For advanced features
# xgboost==2.0.3
# shap==0.44.1
# redis==5.0.1
# Flask-Caching==2.1.0